--- a/services/ui-src/package.json
+++ b/services/ui-src/package.json
@@ -169,6 +169,7 @@
     "eslint-config-prettier": "^7.1.0",
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-prettier": "^3.3.1",
+    "history": "^4.9.0",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",