--- a/services/ui-src/setupEnzyme.js
+++ b/services/ui-src/setupEnzyme.js
@@ -1,4 +1,0 @@
-import "@testing-library/jest-dom/extend-expect";
-import { configure } from "enzyme";
-import Adapter from "enzyme-adapter-react-16";
-configure({ adapter: new Adapter() });