… services/ui-src/config/jest/cssTransform.js
… services/ui-src/config/jest/fileTransform.js
… services/ui-src/src/components/FormFooter/FormFooter.test.jsx
… services/ui-src/src/components/FormHeader/FormHeader.scss
… services/ui-src/src/components/FormHeader/FormHeader.test.jsx
… services/ui-src/src/components/FormPage/FormPage.test.jsx
… services/ui-src/src/components/FormTemplates/FormTemplates.test.jsx
… services/ui-src/src/components/GREGridWithTotals/GREGridWithTotals.test.jsx
… services/ui-src/src/components/GenerateForms/GenerateForms.test.jsx
… services/ui-src/src/components/GenerateTotals/GenerateTotals.test.jsx
… services/ui-src/src/components/GridWithTotals/GridWithTotals.test.jsx
… services/ui-src/src/components/Home/Home.test.jsx
… services/ui-src/src/components/HomeAdmin/HomeAdmin.test.jsx
… services/ui-src/src/components/HomeState/HomeState.test.jsx
… services/ui-src/src/components/Login/Login.test.jsx
… services/ui-src/src/components/NotApplicable/NotApplicable.test.jsx
… services/ui-src/src/components/NotFound/NotFound.test.jsx
… services/ui-src/src/components/Preloader/Preloader.test.jsx
… services/ui-src/src/components/Print/PrintPDF.test.jsx
… services/ui-src/src/components/Profile/Profile.test.jsx
… services/ui-src/src/components/Quarterly/Quarterly.test.jsx
… services/ui-src/src/components/Question/Question.test.jsx
… services/ui-src/src/components/Signup/Signup.test.jsx
… services/ui-src/src/components/StateSelector/StateSelector.test.jsx
… services/ui-src/src/components/SummaryNotes/SummaryNotes.test.jsx
… services/ui-src/src/components/SummaryTab/SummaryTab.test.jsx
… services/ui-src/src/components/SynthesizedGrid/SynthesizedGrid.test.jsx
… services/ui-src/src/components/SynthesizedGridSummary/SynthesizedGridSummary.test.js
… services/ui-src/src/components/SynthesizedGridSummary/SynthesizedGridSummary.test.jsx
… services/ui-src/src/components/TabContainer/TabContainer.test.jsx