@rocio-desantiago 2024-07-31
@benmartin-coforma Correct. 
Actual branch results are even faster by at least half a second!
Test Suites: 41 passed, 41 total
Tests:       1 skipped, 139 passed, 140 total
Snapshots:   0 total
Time:        8.166 s
Ran all test suites.
✨  Done in 8.86s.